# Implement PoT proving and verification optimized for AES (aarch64)