--- a/.github/workflows/rust.yml
+++ b/.github/workflows/rust.yml
@@ -255,14 +255,16 @@ jobs:
       - name: Ensure no panics in annotated code
         env:
           # Increase inlining threshold to make sure the compiler can see that some functions do not panic
-          RUSTFLAGS: ${{ env.RUSTFLAGS }} -Cllvm-args=--inline-threshold=2000
+          # Native CPU and LTO to allow compiler to apply more optimizations and prove lack of panics in more cases
+          RUSTFLAGS: ${{ env.RUSTFLAGS }} -Cllvm-args=--inline-threshold=3000 -C embed-bitcode -C lto=fat -Z dylib-lto -C target-cpu=native
         run: |
           cargo -Zgitoxide -Zgit build --release --all-targets --features no-panic
 
       - name: Ensure no panics in annotated code (various features)
         env:
           # Increase inlining threshold to make sure the compiler can see that some functions do not panic
-          RUSTFLAGS: ${{ env.RUSTFLAGS }} -Cllvm-args=--inline-threshold=2000
+          # Native CPU and LTO to allow compiler to apply more optimizations and prove lack of panics in more cases
+          RUSTFLAGS: ${{ env.RUSTFLAGS }} -Cllvm-args=--inline-threshold=3000 -C embed-bitcode -C lto=fat -Z dylib-lto -C target-cpu=native
         run: |
           # Ensure no panics with `guest` feature
           echo "Checking `no-panic` in contracts"