--- a/subspace/docker/farmer.Dockerfile
+++ b/subspace/docker/farmer.Dockerfile
@@ -106,7 +106,7 @@ RUN \
     ; fi && \
     if [ $TARGETARCH = "amd64" ] && [ "$RUSTFLAGS" = "" ]; then \
       case "$TARGETVARIANT" in \
-        # x86-64-v2 with AES-NI
+        # x86-64-v2
         "v2") export RUSTFLAGS="-C target-cpu=x86-64-v2" ;; \
         # x86-64-v3 with AES-NI
         "v3") export RUSTFLAGS="-C target-cpu=x86-64-v3 -C target-feature=+aes" ;; \