This implements optimized implementation of PoT proving and verification on aarch64.

Before (Raspberry PI 5):
```
verify                  time:   [1.1665 s 1.1817 s 1.1976 s]
```

After:
```
verify                  time:   [835.30 ms 835.30 ms 835.31 ms]
```

I'm not sure why runtime detection is so expensive in `aes` crate (https://github.com/RustCrypto/block-ciphers/issues/484), but we certainly avoid doing the check on each iteration by implementing an efficient implementation + saving some on extra XORs. When compiled with `RUSTFLAGS="-C target-feature=+aes"` `aes` crate improves, but not quite to the same level as optimized implementation, so this is a substantial win either way for generic aarch64 platforms (not sure about the impact on macOS, I think `+aes` is implied there by default):
```
verify                  time:   [866.87 ms 884.87 ms 902.90 ms]
```

Moreover, runtime feature detection now happens once per 200M iterations instead of on each iteration, massively decreasing its cost.

Proving is not expected to be used on aarch64, but for those who do it will be faster now too.